requirements.txt
annotated-types==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
bcrypt==4.2.0 ; python_version >= "3.10" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.17.0 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
cryptography==43.0.3 ; python_version >= "3.10" and python_version < "4.0"
file-retriever @ git+https://github.com/BookOps-CAT/file-retriever.git@cf51d2cbb8c83aa9019f8dceae0f7ae918547525 ; python_version >= "3.10" and python_version < "4.0"
google-api-core==2.20.0 ; python_version >= "3.10" and python_version < "4.0"
google-api-python-client==2.146.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth-oauthlib==1.2.1 ; python_version >= "3.10" and python_version < "4.0"
google-auth==2.35.0 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.65.0 ; python_version >= "3.10" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.10 ; python_version >= "3.10" and python_version < "4.0"
loggly-python-handler==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
numpy==2.1.1 ; python_version >= "3.10" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.10" and python_version < "4.0"
pandas-stubs==2.2.2.240909 ; python_version >= "3.10" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.10" and python_version < "4.0"
paramiko==3.4.1 ; python_version >= "3.10" and python_version < "4.0"
proto-plus==1.24.0 ; python_version >= "3.10" and python_version < "4.0"
protobuf==5.28.2 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.10" and python_version < "4.0"
pydantic-core==2.23.3 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.9.1 ; python_version >= "3.10" and python_version < "4.0"
pymarc==5.2.2 ; python_version >= "3.10" and python_version < "4.0"
pynacl==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.1.4 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
pytz==2024.2 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.10" and python_version < "4.0"
record-validator @ git+https://github.com/BookOps-CAT/record-validator.git@76efaaebf79fe8f1a6675720339915a8dc8e5853 ; python_version >= "3.10" and python_version < "4.0"
requests-futures==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
requests-oauthlib==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.10" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
types-paramiko==3.4.0.20240423 ; python_version >= "3.10" and python_version < "4.0"
types-pytz==2024.2.0.20241003 ; python_version >= "3.10" and python_version < "4.0"
types-pyyaml==6.0.12.20240808 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2024.2 ; python_version >= "3.10" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.10" and python_version < "4.0"
